moved the signup.js out of dir/auth, changed sing up button on login on signup.js; changed register button on login.js to say sign up
