+++ plant-app/src/App.js
@@ -2,9 +2,8 @@ import React from 'react';
 import './App.css';
 
 import { BrowserRouter as Router, Route, Switch} from 'react-router-dom';
-import PrivateRoute from './Auth/PrivateRoute';
 
-import SignUp from './Auth/SignUp';
+import SignUp from './components/SignUp';
 import Login from './components/Login'
