+++ plant-app/src/components/SignUp.js
@@ -1,5 +1,5 @@
-import React, { useState } from 'react'
-import { axiosWithAuth } from './axiosWithAuth'
+import React, { useState } from 'react';
+import { axiosWithAuth } from '../auth/axiosWithAuth';
 import {BrowserRouter as Router, Route, Link, Switch } from 'react-router-dom';
 
 import styled from 'styled-components'
@@ -40,7 +40,7 @@ const SignUp = (props) => {
  
         <div>
         <h1>Welcome!</h1>
-        <p>Register your account below</p>
+        <p>Sign up for an account below</p>
           <form onSubmit={signUp}>
           <input
             type='text'
@@ -56,10 +56,10 @@ const SignUp = (props) => {
             value={credentials.password}
             onChange={handleChange}
           />
-          <button onClick={signUp}>Register</button>
+          <button onClick={signUp}>Sign Up!</button>
           </form>
                   
-          <Link to='/login'>Already have an account? - Sign In</Link>
+          <Link to='/login'>Already have an account? - Login</Link>
        
         </div>
